refactor: Move transaction details dialog to App component
